fixed many MSVC warnings
git-svn-id: file://localhost/home/tolik/svnrepo/trunk/opencv@2004 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ 3rdparty/flann/algorithms/kdtree_index.h
@@ -596,7 +596,7 @@ private:
 			adding exceeds their value.
 		*/
 
-		double new_distsq = flann_dist(&val, &val+1, &node->divval, mindistsq);
+		float new_distsq = flann_dist(&val, &val+1, &node->divval, mindistsq);
 //		if (2 * checkCount < maxCheck  ||  !result.full()) {
 		if (new_distsq < result.worstDist() ||  !result.full()) {
 			heap->insert( BranchSt::make_branch(otherChild, new_distsq) );
@@ -642,7 +642,7 @@ private:
 
 		/* Call recursively to search next level down. */
 		searchLevelExact(result, vec, bestChild, mindistsq);
-		double new_distsq = flann_dist(&val, &val+1, &node->divval, mindistsq);
+		float new_distsq = flann_dist(&val, &val+1, &node->divval, mindistsq);
 		searchLevelExact(result, vec, otherChild, new_distsq);
 	}
